CHROMIUM: iwl7000: add wide firmware command infrastructure for TX
As the firmware is slowly running out of command IDs and grouping of
commands is desirable anyway, the firmware is extending the command
header from 4 bytes to 8 bytes to introduce a group (in place of the
former flags field, since that's always 0 on commands and thus can
be easily used to distinguish between the two.

In order to support this most easily in the driver widen the command
command ID used in the command sending functions and encode the new
values (group and version) in the ID. That way existing code doesn't
have to be changed (since the higher bits are 0 automatically) and
newer code can easily use the new ID generation function to create a
value to use in place of just the command ID.

iwl7000-tree: 83ce5ff0843aafbaf263c712c6c7d431b6d4ebb0

+++ drivers/net/wireless-3.8/iwl7000/iwlwifi/xvt/utils.c
@@ -79,7 +79,7 @@ int iwl_xvt_send_cmd(struct iwl_xvt *xvt, struct iwl_host_cmd *cmd)
 	return iwl_trans_send_cmd(xvt->trans, cmd);
 }
 
-int iwl_xvt_send_cmd_pdu(struct iwl_xvt *xvt, u8 id,
+int iwl_xvt_send_cmd_pdu(struct iwl_xvt *xvt, u32 id,
 			 u32 flags, u16 len, const void *data)
 {
 	struct iwl_host_cmd cmd = {
